requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
backoff==1.11.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
multidict==6.1.0
packaging==24.1
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
spotify==0.10.2
spotipy==2.24.0
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
